requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.2
